+++ src/components/ExpenseCreate/ExpenseCreate.js
@@ -1,10 +1,87 @@
 import React from "react";
-import './ExpenseCreate.scss';
-
+import { Box, Button, TextField, Typography } from "@mui/material";
 
 export default function ExpenseCreate(props) {
+  const handleSubmit = () => {
+    return null;
+  };
 
   return (
+    <Box
+      sx={{
+        marginTop: 8,
+        display: "flex",
+        flexDirection: "column",
+        alignItems: "center",
+      }}
+    >
+      <Typography sx={{ fontFamily: "monospace" }}>Add New</Typography>
+      <Box component="form" onSubmit={handleSubmit} noValidate sx={{ mt: 1 }}>
+        <TextField
+          margin="normal"
+          fullWidth
+          id="cost"
+          label="Price of item"
+          name="cost"
+          autoFocus
+        />
+        <TextField
+          margin="normal"
+          fullWidth
+          id="currency"
+          label="Currency of Purchase"
+          name="currencyid"
+        />
+        <TextField
+          margin="normal"
+          required
+          fullWidth
+          id="date"
+          name="datePaid"
+          type="date"
+        />
+        <TextField
+          margin="normal"
+          fullWidth
+          id="notes"
+          label="Notes"
+          name="notes"
+        />
+
+        <Button
+          type="submit"
+          fullWidth
+          variant="contained"
+          sx={{
+            mt: 3,
+            mb: 2,
+            fontFamily: "monospace",
+            bgcolor: "#6D89AE",
+            "&:hover": { bgcolor: "#9ACCE3" },
+          }}
+        >
+          Submit
+        </Button>
+        {/*         <Grid container>
+              {error.active && (
+                <Grid
+                  item
+                  sx={{
+                    fontFamily: "monospace",
+                    my: 3,
+                    color: "red",
+                    fontSize: "1.2rem",
+                  }}
+                >
+                  {error.msg}
+                </Grid>
+              )}
+            </Grid> */}
+      </Box>
+    </Box>
+  );
+}
+/*  return (
     <div className="form">
       <form onSubmit={(e) => handleSubmit(e)}>
         <div className="input-container">
@@ -57,4 +134,4 @@ export default function ExpenseCreate(props) {
       </form>
     </div>
   );
-}
+} */
